+++ bin/builder.js
@@ -1,5 +1,7 @@
+const childStags = require('../dist/animations/childStag').childStags;
+
 exports.fxArrayToImportSyntax = function fxArrayToImportSyntax(fxArray){
-  return 'import { AnimationTriggerMetadata,trigger,style,state,transition,animate,keyframes } from \'@angular/animations\'\n'+
+  return 'import { query, stagger, animateChild, AnimationTriggerMetadata,trigger,style,state,transition,animate,keyframes } from \'@angular/animations\'\n'+
   'export const fxArray = ['+
   fxArrayToSyntaxArray(fxArray).join() + ']'
 }
@@ -10,21 +12,15 @@ function fxArrayToSyntaxArray(fxArray){
   const triggers = []
   for(let trigIndex=fxArray.length-1; trigIndex >= 0; --trigIndex){
     let fx = fxArray[trigIndex]
-    let states = pushStateDefs(fx.definitions)
+    let states = pushStateDefs(fx.definitions,fx)
     let name = fx.name
-    //let trig = trigger(name, states)
-    //animations.push( trig )
-    
-    //triggers.push(`\npublic static _${name} = trigger('${name}', ${states})`)
+
+    if(fx.name=='childStag'){
+      triggers.push.apply(triggers, childStags.map(trig=>triggerToString(trig)))
+      continue;
+    }
+
     triggers.push(`\ntrigger('${name}', ${states})`)
-    
-    /*
-    if(array[trigIndex].name == 'absoluteSwap500'){
-      console.log('array[trigIndex].name',array[trigIndex].name)
-      console.log('states0', states[0])
-      //console.log('states', states)
-      console.log('trig', trig)
-    }*/
   }
 
   return triggers
@@ -46,9 +42,10 @@ function pushStateDefs(defs){
       style = JSON.stringify( def.styles.styles || def.styles.styles[0] )//Angular v4.0.0+ || v2.4.0+
       steps = `\nstyle(${style})`
       trans = `\nstate('${expr}', ${steps})`
+    }else if(def.definitions){
+      trans = triggerToString(def)//Angular v4.0.0+ || v2.4.0+
     }else{
       steps = animateSteps(def.animation || def.steps._steps)//Angular v4.0.0+ || v2.4.0+
-      //let trans = transition(def["stateChangeExpr"], steps)
       trans = `\ntransition('${expr}', ${steps})`
     }
     
@@ -63,17 +60,61 @@ function pushStateDefs(defs){
 */
 function animateSteps(steps){
   const animations = []
-//console.log('steps.styles.steps.length', steps)
   
   const stepCount = steps.length
   for(let stepIndex=0; stepIndex < stepCount; ++stepIndex){
     let step = steps[stepIndex]
-    let keySteps = keyframeSteps(step.styles.steps)
-    //let anim = animate(step.timings, keySteps)
-    let anim = `\nanimate('${step.timings}', ${keySteps})`
+    let anim = animationToString(step)
+
     animations.push( anim )
   }
   return '['+animations.join()+']'
+}
+
+function animationToString(step){
+  if( step.selector ){
+    return queryToString(step)
+  }else if( isStaggerStep(step) ){
+    return staggerToString(step)
+  }else if( isAnimateChild(step) ){
+    return animateChildToString(step)
+  }else{
+    return animateToString(step)
+  }
+}
+
+function animateToString(step){
+  let keySteps = keyframeSteps(step.styles.steps)
+  //anim = animate(step.timings, keySteps)
+  return `\nanimate('${step.timings}', ${keySteps})`
+}
+
+function triggerToString(step){
+  return `\ntrigger('${step.name}', ${pushStateDefs(step.definitions)})`
+}
+
+function animateChildToString(step){
+  return `\nanimateChild()`
+}
+
+function intQuoter(v){
+  return isNaN(Number(v)) ? "'"+v+"'" : Number(v)
+}
+
+function staggerToString(step){
+  return `\nstagger(${intQuoter(step.timings)}, ${animateSteps(step.animation)})`
+}
+
+function isStaggerStep(step){
+  return step.type==12
+}
+
+function isAnimateChild(step){
+  return step.type==9
+}
+
+function queryToString(step){
+  return `\nquery('${step.selector}', ${animateSteps(step.animation)}, ${JSON.stringify(step.options,null,2)})`
 }
 
 function keyframeSteps(steps){