+++ src/animations/slide.ts
@@ -1,11 +1,4 @@
-/*import {
-  style,
-  state,
-  transition,
-  animate,
-  keyframes,
-  AnimationMetadata
-} from '@angular/core';*/
+/* NOTE: All the visibility definitions are to ensure staggering works correctly */
 import { keyframes, AnimationMetadata, state, style, transition, animate } from '@angular/animations';
 
 import { defaultOptions, stylize, combo } from "./helper"
@@ -30,50 +23,74 @@ export function slideOptions(options): AnimationMetadata[]{
     })),
     transition(options.options.igniter+' => slideInDown', [
       animate(options.timing, keyframes([
-        stylize({transform: 'translate3d(0, -100%, 0)', offset: 0}, options.options),
-        stylize({transform: 'translate3d(0, 0, 0)', offset: 1}, options.options)
+        stylize({visibility: 'hidden', transform: 'translate3d(0, -100%, 0)', offset: 0}, options.options),
+        stylize({visibility: 'visible', transform: 'translate3d(0, 0, 0)', offset: 1}, options.options)
       ]))
     ]),
-    transition('slideInDown => void, '+options.options.igniter+' => slideOutDown', [
+    transition('slideInDown => void', [
       animate(options.timing, keyframes([
-        stylize({transform: 'translate3d(0, 0, 0)', offset: 0}, options.options),
-        stylize({transform: 'translate3d(0, 100%, 0)', offset: 1}, options.options)
+        stylize({visibility: 'visible', transform: 'translate3d(0, 0, 0)', offset: 0}, options.options),
+        stylize({visibility: 'hidden', transform: 'translate3d(0, 100%, 0)', offset: 1}, options.options)
+      ]))
+    ]),
+    transition(options.options.igniter+' => slideOutDown', [
+      animate(options.timing, keyframes([
+        stylize({visibility: 'hidden', transform: 'translate3d(0, 0, 0)', offset: 0}, options.options),
+        stylize({visibility: 'visible', transform: 'translate3d(0, 100%, 0)', offset: 1}, options.options)
       ]))
     ]),
     transition(options.options.igniter+' => slideInLeft', [
       animate(options.timing, keyframes([
-        stylize({transform: 'translate3d(-100%, 0, 0)', offset: 0}, options.options),
-        stylize({transform: 'translate3d(0, 0, 0)', offset: 1}, options.options)
+        stylize({visibility: 'hidden', transform: 'translate3d(-100%, 0, 0)', offset: 0}, options.options),
+        stylize({visibility: 'visible', transform: 'translate3d(0, 0, 0)', offset: 1}, options.options)
       ]))
     ]),
-    transition('slideInLeft => void, '+options.options.igniter+' => slideOutRight', [
+    transition('slideInLeft => void', [
       animate(options.timing, keyframes([
-        stylize({transform: 'translate3d(0, 0, 0)', offset: 0}, options.options),
-        stylize({transform: 'translate3d(100%, 0, 0)', offset: 1}, options.options)
+        stylize({visibility: 'visible', transform: 'translate3d(0, 0, 0)', offset: 0}, options.options),
+        stylize({visibility: 'hidden', transform: 'translate3d(100%, 0, 0)', offset: 1}, options.options)
+      ]))
+    ]),
+    transition(options.options.igniter+' => slideOutRight', [
+      animate(options.timing, keyframes([
+        stylize({visibility: 'hidden', transform: 'translate3d(0, 0, 0)', offset: 0}, options.options),
+        stylize({visibility: 'visible', transform: 'translate3d(100%, 0, 0)', offset: 1}, options.options)
       ]))
     ]),
     transition(options.options.igniter+' => slideInRight', [
       animate(options.timing, keyframes([
-        stylize({transform: 'translate3d(100%, 0, 0)', offset: 0}, options.options),
-        stylize({transform: 'translate3d(0, 0, 0)', offset: 1}, options.options)
+        stylize({visibility: 'hidden', transform: 'translate3d(100%, 0, 0)', offset: 0}, options.options),
+        stylize({visibility: 'visible', transform: 'translate3d(0, 0, 0)', offset: 1}, options.options)
       ]))
     ]),
-    transition('slideInRight => void, '+options.options.igniter+' => slideOutLeft', [
+    transition('slideInRight => void', [
       animate(options.timing, keyframes([
-        stylize({transform: 'translate3d(0, 0, 0)', offset: 0}, options.options),
-        stylize({transform: 'translate3d(-100%, 0, 0)', offset: 1}, options.options)
+        stylize({visibility: 'visible', transform: 'translate3d(0, 0, 0)', offset: 0}, options.options),
+        stylize({visibility: 'hidden', transform: 'translate3d(-100%, 0, 0)', offset: 1}, options.options)
+      ]))
+    ]),
+    transition(options.options.igniter+' => slideOutLeft', [
+      animate(options.timing, keyframes([
+        stylize({visibility: 'hidden', transform: 'translate3d(0, 0, 0)', offset: 0}, options.options),
+        stylize({visibility: 'visible', transform: 'translate3d(-100%, 0, 0)', offset: 1}, options.options)
       ]))
     ]),
     transition(options.options.igniter+' => slideInUp', [
       animate(options.timing, keyframes([
-        stylize({transform: 'translate3d(0, 100%, 0)', offset: 0}, options.options),
-        stylize({transform: 'translate3d(0, 0, 0)', offset: 1}, options.options)
+        stylize({visibility: 'hidden', transform: 'translate3d(0, 100%, 0)', offset: 0}, options.options),
+        stylize({visibility: 'visible', transform: 'translate3d(0, 0, 0)', offset: 1}, options.options)
       ]))
     ]),
-    transition('slideInUp => void, '+options.options.igniter+' => slideOutUp', [
+    transition('slideInUp => void', [
       animate(options.timing, keyframes([
-        stylize({transform: 'translate3d(0, 0, 0)', offset: 0}, options.options),
-        stylize({transform: 'translate3d(0, -100%, 0)', offset: 1}, options.options)
+        stylize({visibility: 'visible', transform: 'translate3d(0, 0, 0)', offset: 0}, options.options),
+        stylize({visibility: 'hidden', transform: 'translate3d(0, -100%, 0)', offset: 1}, options.options)
+      ]))
+    ]),
+    transition(options.options.igniter+' => slideOutUp', [
+      animate(options.timing, keyframes([
+        stylize({visibility: 'hidden', transform: 'translate3d(0, 0, 0)', offset: 0}, options.options),
+        stylize({visibility: 'visible', transform: 'translate3d(0, -100%, 0)', offset: 1}, options.options)
       ]))
     ])
   ]